// frontend/src/app/models/korisnik.ts
export class Korisnik{
  korisnicko_ime: string = "";
  lozinka: string = "";
  bezbednosno_pitanje: string = "";
  odgovor_na_bezbednosno_pitanje: string = "";
  ime: string = "";
  prezime: string = "";
  pol: string = "";
  adresa: string = "";
  kontakt_telefon: string = "";
  email: string = "";
  profilna_slika: string = "";
  broj_kreditne_kartice: string = "";
  tip: string = "";
  odobren: boolean = false;
  idRestorana: number = 0;
  restoran: string = "";
  bodovi: number = 0;
  blokiran: boolean = false;
  deaktiviran: boolean = false;
  nijeOdobren: boolean = false;
}